handle alt id
when an auth0 user being imported into fusionauth has previously been imported into auth0, you are not guaranteed a uuid in the auth0 id field.

Instead, you have the previous uid (which could be an int) stored in the user id field. The secrets file also has an alt_id field, but also contains an auth0 uuid in the  field. So we use that for our FusionAuth id when we map previous ids.